Comments! Comments Everywhere!
So basically this is task2

Main2 takes an argument

If argument not a file, or missing - display message.

If argument is a file, it has to checkout accordingly
to some content design (see iodesign readonly.txt)

If the design of the given file is not right, then you're promted
if you want to re-initialise it to proper content design for this
project.

Every output is appended to the same file.

Checkout iodesign readonly.txt and you'll get the ideea.

Commented as much as i could on this one.

bulletproof so far (based on some specified limitations).